// server/src/modules/combat/battle-rules.generated.ts
export const COMBAT_RANGE_KEYS = ["physicalAttack",
  "magicAttack",
  "taoistAttack",
  "physicalDefense",
  "magicDefense",
  "taoistDefense"];
// ...
export const COMBAT_WEIGHTS = {
  "magicAttack": 5,
  "physicalAttack": 5,
  "taoistAttack": 5,
  "magicDefense": 10,
  "physicalDefense": 10,
  "taoistDefense": 10,
  "maxHp": 15,
};